+++ spring-blog-ang/src/app/auth.service.ts
@@ -32,4 +32,9 @@ export class AuthService {
   isAuthenticated(): boolean {
     return this.localStorageService.retrieve('username') != null;
   }
+
+  logout() {
+    this.localStorageService.clear("authenticationToken");
+    this.localStorageService.clear("username");
+  }
 }